+++ .gitignore
@@ -2,3 +2,4 @@
 **/*.rs.bk
 Cargo.lock
 .idea
+*.env
