+++ .rustfmt.toml
@@ -1,0 +1,1 @@
+edition = "2021"
