+++ src/s3/client.rs
@@ -23,13 +23,14 @@ use crate::s3::response::*;
 use crate::s3::signer::{presign_v4, sign_v4_s3};
 use crate::s3::sse::SseCustomerKey;
 use crate::s3::types::{
-    Bucket, DeleteObject, Directive, Item, LifecycleConfig, NotificationConfig, ObjectLockConfig,
-    Part, ReplicationConfig, RetentionMode, SseConfig,
+    Bucket, DeleteObject, Directive, LifecycleConfig, NotificationConfig, ObjectLockConfig, Part,
+    ReplicationConfig, RetentionMode, SseConfig,
 };
 use crate::s3::utils::{
     from_iso8601utc, get_default_text, get_option_text, get_text, md5sum_hash, merge, sha256_hash,
-    to_amz_date, to_iso8601utc, urldecode, utc_now, Multimap,
+    to_amz_date, to_iso8601utc, utc_now, Multimap,
 };
+
 use async_recursion::async_recursion;
 use bytes::{Buf, Bytes};
 use dashmap::DashMap;
@@ -44,169 +45,11 @@ use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use xmltree::Element;
 
+mod list_objects;
 mod listen_bucket_notification;
 
+pub use list_objects::*;
 pub use listen_bucket_notification::*;
-
-fn url_decode(
-    encoding_type: &Option<String>,
-    prefix: Option<String>,
-) -> Result<Option<String>, Error> {
-    if let Some(v) = encoding_type.as_ref() {
-        if v == "url" {
-            if let Some(v) = prefix {
-                return Ok(Some(urldecode(&v)?.to_string()));
-            }
-        }
-    }
-
-    if let Some(v) = prefix.as_ref() {
-        return Ok(Some(v.to_string()));
-    }
-
-    Ok(None)
-}
-
-fn add_common_list_objects_query_params(
-    query_params: &mut Multimap,
-    delimiter: Option<&str>,
-    encoding_type: Option<&str>,
-    max_keys: Option<u16>,
-    prefix: Option<&str>,
-) {
-    query_params.insert(
-        String::from("delimiter"),
-        delimiter.unwrap_or("").to_string(),
-    );
-    query_params.insert(
-        String::from("max-keys"),
-        max_keys.unwrap_or(1000).to_string(),
-    );
-    query_params.insert(String::from("prefix"), prefix.unwrap_or("").to_string());
-    if let Some(v) = encoding_type {
-        query_params.insert(String::from("encoding-type"), v.to_string());
-    }
-}
-
-fn parse_common_list_objects_response(
-    root: &Element,
-) -> Result<
-    (
-        String,
-        Option<String>,
-        Option<String>,
-        Option<String>,
-        bool,
-        Option<u16>,
-    ),
-    Error,
-> {
-    let encoding_type = get_option_text(root, "EncodingType");
-    let prefix = url_decode(&encoding_type, Some(get_default_text(root, "Prefix")))?;
-    Ok((
-        get_text(root, "Name")?,
-        encoding_type,
-        prefix,
-        get_option_text(root, "Delimiter"),
-        match get_option_text(root, "IsTruncated") {
-            Some(v) => v.to_lowercase() == "true",
-            None => false,
-        },
-        match get_option_text(root, "MaxKeys") {
-            Some(v) => Some(v.parse::<u16>()?),
-            None => None,
-        },
-    ))
-}
-
-fn parse_list_objects_contents(
-    contents: &mut Vec<Item>,
-    root: &mut xmltree::Element,
-    tag: &str,
-    encoding_type: &Option<String>,
-    is_delete_marker: bool,
-) -> Result<(), Error> {
-    while let Some(v) = root.take_child(tag) {
-        let content = v;
-        let etype = encoding_type.as_ref().cloned();
-        let key = url_decode(&etype, Some(get_text(&content, "Key")?))?.unwrap();
-        let last_modified = Some(from_iso8601utc(&get_text(&content, "LastModified")?)?);
-        let etag = get_option_text(&content, "ETag");
-        let v = get_default_text(&content, "Size");
-        let size = match v.is_empty() {
-            true => None,
-            false => Some(v.parse::<usize>()?),
-        };
-        let storage_class = get_option_text(&content, "StorageClass");
-        let is_latest = get_default_text(&content, "IsLatest").to_lowercase() == "true";
-        let version_id = get_option_text(&content, "VersionId");
-        let (owner_id, owner_name) = match content.get_child("Owner") {
-            Some(v) => (get_option_text(v, "ID"), get_option_text(v, "DisplayName")),
-            None => (None, None),
-        };
-        let user_metadata = match content.get_child("UserMetadata") {
-            Some(v) => {
-                let mut map: HashMap<String, String> = HashMap::new();
-                for xml_node in &v.children {
-                    let u = xml_node
-                        .as_element()
-                        .ok_or(Error::XmlError("unable to convert to element".to_string()))?;
-                    map.insert(
-                        u.name.to_string(),
-                        u.get_text().unwrap_or_default().to_string(),
-                    );
-                }
-                Some(map)
-            }
-            None => None,
-        };
-
-        contents.push(Item {
-            name: key,
-            last_modified,
-            etag,
-            owner_id,
-            owner_name,
-            size,
-            storage_class,
-            is_latest,
-            version_id,
-            user_metadata,
-            is_prefix: false,
-            is_delete_marker,
-            encoding_type: etype,
-        });
-    }
-
-    Ok(())
-}
-
-fn parse_list_objects_common_prefixes(
-    contents: &mut Vec<Item>,
-    root: &mut Element,
-    encoding_type: &Option<String>,
-) -> Result<(), Error> {
-    while let Some(v) = root.take_child("CommonPrefixes") {
-        let common_prefix = v;
-        contents.push(Item {
-            name: url_decode(encoding_type, Some(get_text(&common_prefix, "Prefix")?))?.unwrap(),
-            last_modified: None,
-            etag: None,
-            owner_id: None,
-            owner_name: None,
-            size: None,
-            storage_class: None,
-            is_latest: false,
-            version_id: None,
-            user_metadata: None,
-            is_prefix: true,
-            is_delete_marker: false,
-            encoding_type: encoding_type.as_ref().cloned(),
-        });
-    }
-
-    Ok(())
-}
 
 /// Client Builder manufactures a Client using given parameters.
 #[derive(Debug, Default)]
@@ -315,7 +158,7 @@ impl Client {
     /// use minio::s3::client::Client;
     /// use minio::s3::creds::StaticProvider;
     /// use minio::s3::http::BaseUrl;
-    /// let mut base_url: BaseUrl = "play.min.io".parse().unwrap();
+    /// let base_url: BaseUrl = "play.min.io".parse().unwrap();
     /// let static_provider = StaticProvider::new(
     ///     "Q3AM3UQ867SPQQA43P2F",
     ///     "zuf+tfteSlswRu7BJ86wekitnifILbZam1KYY3TG",
@@ -2431,327 +2274,6 @@ impl Client {
             headers: header_map.clone(),
             buckets: bucket_list,
         })
-    }
-
-    /// Executes [ListObjects](https://docs.aws.amazon.com/AmazonS3/latest/API/API_ListObjects.html) S3 API
-    pub async fn list_objects_v1(
-        &self,
-        args: &ListObjectsV1Args<'_>,
-    ) -> Result<ListObjectsV1Response, Error> {
-        let region = self.get_region(args.bucket, args.region).await?;
-
-        let mut headers = Multimap::new();
-        if let Some(v) = &args.extra_headers {
-            merge(&mut headers, v);
-        }
-
-        let mut query_params = Multimap::new();
-        if let Some(v) = &args.extra_query_params {
-            merge(&mut query_params, v);
-        }
-        add_common_list_objects_query_params(
-            &mut query_params,
-            args.delimiter,
-            args.encoding_type,
-            args.max_keys,
-            args.prefix,
-        );
-        if let Some(v) = &args.marker {
-            query_params.insert(String::from("marker"), v.to_string());
-        }
-
-        let resp = self
-            .execute(
-                Method::GET,
-                &region,
-                &mut headers,
-                &query_params,
-                Some(args.bucket),
-                None,
-                None,
-            )
-            .await?;
-        let header_map = resp.headers().clone();
-        let body = resp.bytes().await?;
-        let mut root = Element::parse(body.reader())?;
-
-        let (name, encoding_type, prefix, delimiter, is_truncated, max_keys) =
-            parse_common_list_objects_response(&root)?;
-        let marker = url_decode(&encoding_type, get_option_text(&root, "Marker"))?;
-        let mut next_marker = url_decode(&encoding_type, get_option_text(&root, "NextMarker"))?;
-        let mut contents: Vec<Item> = Vec::new();
-        parse_list_objects_contents(&mut contents, &mut root, "Contents", &encoding_type, false)?;
-        if is_truncated && next_marker.is_none() {
-            next_marker = contents.last().map(|v| v.name.clone())
-        }
-        parse_list_objects_common_prefixes(&mut contents, &mut root, &encoding_type)?;
-
-        Ok(ListObjectsV1Response {
-            headers: header_map,
-            name,
-            encoding_type,
-            prefix,
-            delimiter,
-            is_truncated,
-            max_keys,
-            contents,
-            marker,
-            next_marker,
-        })
-    }
-
-    /// Executes [ListObjectsV2](https://docs.aws.amazon.com/AmazonS3/latest/API/API_ListObjectsV2.html) S3 API
-    pub async fn list_objects_v2(
-        &self,
-        args: &ListObjectsV2Args<'_>,
-    ) -> Result<ListObjectsV2Response, Error> {
-        let region = self.get_region(args.bucket, args.region).await?;
-
-        let mut headers = Multimap::new();
-        if let Some(v) = &args.extra_headers {
-            merge(&mut headers, v);
-        }
-
-        let mut query_params = Multimap::new();
-        if let Some(v) = &args.extra_query_params {
-            merge(&mut query_params, v);
-        }
-        query_params.insert(String::from("list-type"), String::from("2"));
-        add_common_list_objects_query_params(
-            &mut query_params,
-            args.delimiter,
-            args.encoding_type,
-            args.max_keys,
-            args.prefix,
-        );
-        if let Some(v) = &args.continuation_token {
-            query_params.insert(String::from("continuation-token"), v.to_string());
-        }
-        if args.fetch_owner {
-            query_params.insert(String::from("fetch-owner"), String::from("true"));
-        }
-        if let Some(v) = &args.start_after {
-            query_params.insert(String::from("start-after"), v.to_string());
-        }
-        if args.include_user_metadata {
-            query_params.insert(String::from("metadata"), String::from("true"));
-        }
-
-        let resp = self
-            .execute(
-                Method::GET,
-                &region,
-                &mut headers,
-                &query_params,
-                Some(args.bucket),
-                None,
-                None,
-            )
-            .await?;
-        let header_map = resp.headers().clone();
-        let body = resp.bytes().await?;
-        let mut root = Element::parse(body.reader())?;
-
-        let (name, encoding_type, prefix, delimiter, is_truncated, max_keys) =
-            parse_common_list_objects_response(&root)?;
-        let text = get_option_text(&root, "KeyCount");
-        let key_count = match text {
-            Some(v) => match v.is_empty() {
-                true => None,
-                false => Some(v.parse::<u16>()?),
-            },
-            None => None,
-        };
-        let start_after = url_decode(&encoding_type, get_option_text(&root, "StartAfter"))?;
-        let continuation_token = get_option_text(&root, "ContinuationToken");
-        let next_continuation_token = get_option_text(&root, "NextContinuationToken");
-        let mut contents: Vec<Item> = Vec::new();
-        parse_list_objects_contents(&mut contents, &mut root, "Contents", &encoding_type, false)?;
-        parse_list_objects_common_prefixes(&mut contents, &mut root, &encoding_type)?;
-
-        Ok(ListObjectsV2Response {
-            headers: header_map,
-            name,
-            encoding_type,
-            prefix,
-            delimiter,
-            is_truncated,
-            max_keys,
-            contents,
-            key_count,
-            start_after,
-            continuation_token,
-            next_continuation_token,
-        })
-    }
-
-    /// Executes [ListObjectVersions](https://docs.aws.amazon.com/AmazonS3/latest/API/API_ListObjectVersions.html) S3 API
-    pub async fn list_object_versions(
-        &self,
-        args: &ListObjectVersionsArgs<'_>,
-    ) -> Result<ListObjectVersionsResponse, Error> {
-        let region = self.get_region(args.bucket, args.region).await?;
-
-        let mut headers = Multimap::new();
-        if let Some(v) = &args.extra_headers {
-            merge(&mut headers, v);
-        }
-
-        let mut query_params = Multimap::new();
-        if let Some(v) = &args.extra_query_params {
-            merge(&mut query_params, v);
-        }
-        query_params.insert(String::from("versions"), String::new());
-        add_common_list_objects_query_params(
-            &mut query_params,
-            args.delimiter,
-            args.encoding_type,
-            args.max_keys,
-            args.prefix,
-        );
-        if let Some(v) = &args.key_marker {
-            query_params.insert(String::from("key-marker"), v.to_string());
-        }
-        if let Some(v) = &args.version_id_marker {
-            query_params.insert(String::from("version-id-marker"), v.to_string());
-        }
-
-        let resp = self
-            .execute(
-                Method::GET,
-                &region,
-                &mut headers,
-                &query_params,
-                Some(args.bucket),
-                None,
-                None,
-            )
-            .await?;
-        let header_map = resp.headers().clone();
-        let body = resp.bytes().await?;
-        let mut root = Element::parse(body.reader())?;
-
-        let (name, encoding_type, prefix, delimiter, is_truncated, max_keys) =
-            parse_common_list_objects_response(&root)?;
-        let key_marker = url_decode(&encoding_type, get_option_text(&root, "KeyMarker"))?;
-        let next_key_marker = url_decode(&encoding_type, get_option_text(&root, "NextKeyMarker"))?;
-        let version_id_marker = get_option_text(&root, "VersionIdMarker");
-        let next_version_id_marker = get_option_text(&root, "NextVersionIdMarker");
-        let mut contents: Vec<Item> = Vec::new();
-        parse_list_objects_contents(&mut contents, &mut root, "Version", &encoding_type, false)?;
-        parse_list_objects_common_prefixes(&mut contents, &mut root, &encoding_type)?;
-        parse_list_objects_contents(
-            &mut contents,
-            &mut root,
-            "DeleteMarker",
-            &encoding_type,
-            true,
-        )?;
-
-        Ok(ListObjectVersionsResponse {
-            headers: header_map,
-            name,
-            encoding_type,
-            prefix,
-            delimiter,
-            is_truncated,
-            max_keys,
-            contents,
-            key_marker,
-            next_key_marker,
-            version_id_marker,
-            next_version_id_marker,
-        })
-    }
-
-    /// List objects with version information optionally. `results_fn` callback
-    /// function is repeatedly called with object information and returning
-    /// false from the callback stops further listing.
-    pub async fn list_objects(&self, args: &ListObjectsArgs<'_>) -> Result<(), Error> {
-        let mut lov1_args = ListObjectsV1Args::new(args.bucket)?;
-        lov1_args.extra_headers = args.extra_headers;
-        lov1_args.extra_query_params = args.extra_query_params;
-        lov1_args.region = args.region;
-        if args.recursive {
-            lov1_args.delimiter = None;
-        } else {
-            lov1_args.delimiter = Some(args.delimiter.unwrap_or("/"));
-        }
-        lov1_args.encoding_type = match args.use_url_encoding_type {
-            true => Some("url"),
-            false => None,
-        };
-        lov1_args.max_keys = args.max_keys;
-        lov1_args.prefix = args.prefix;
-        lov1_args.marker = args.marker.map(|x| x.to_string());
-
-        let mut lov2_args = ListObjectsV2Args::new(args.bucket)?;
-        lov2_args.extra_headers = args.extra_headers;
-        lov2_args.extra_query_params = args.extra_query_params;
-        lov2_args.region = args.region;
-        if args.recursive {
-            lov2_args.delimiter = None;
-        } else {
-            lov2_args.delimiter = Some(args.delimiter.unwrap_or("/"));
-        }
-        lov2_args.encoding_type = match args.use_url_encoding_type {
-            true => Some("url"),
-            false => None,
-        };
-        lov2_args.max_keys = args.max_keys;
-        lov2_args.prefix = args.prefix;
-        lov2_args.start_after = args.start_after.map(|x| x.to_string());
-        lov2_args.continuation_token = args.continuation_token.map(|x| x.to_string());
-        lov2_args.fetch_owner = args.fetch_owner;
-        lov2_args.include_user_metadata = args.include_user_metadata;
-
-        let mut lov_args = ListObjectVersionsArgs::new(args.bucket)?;
-        lov_args.extra_headers = args.extra_headers;
-        lov_args.extra_query_params = args.extra_query_params;
-        lov_args.region = args.region;
-        if args.recursive {
-            lov_args.delimiter = None;
-        } else {
-            lov_args.delimiter = Some(args.delimiter.unwrap_or("/"));
-        }
-        lov_args.encoding_type = match args.use_url_encoding_type {
-            true => Some("url"),
-            false => None,
-        };
-        lov_args.max_keys = args.max_keys;
-        lov_args.prefix = args.prefix;
-        lov_args.key_marker = args.key_marker.map(|x| x.to_string());
-        lov_args.version_id_marker = args.version_id_marker.map(|x| x.to_string());
-
-        let mut stop = false;
-        while !stop {
-            if args.include_versions {
-                let resp = self.list_object_versions(&lov_args).await?;
-                stop = !resp.is_truncated;
-                if resp.is_truncated {
-                    lov_args.key_marker = resp.next_key_marker;
-                    lov_args.version_id_marker = resp.next_version_id_marker;
-                }
-                stop = stop || !(args.result_fn)(resp.contents);
-            } else if args.use_api_v1 {
-                let resp = self.list_objects_v1(&lov1_args).await?;
-                stop = !resp.is_truncated;
-                if resp.is_truncated {
-                    lov1_args.marker = resp.next_marker;
-                }
-                stop = stop || !(args.result_fn)(resp.contents);
-            } else {
-                let resp = self.list_objects_v2(&lov2_args).await?;
-                stop = !resp.is_truncated;
-                if resp.is_truncated {
-                    lov2_args.start_after = resp.start_after;
-                    lov2_args.continuation_token = resp.next_continuation_token;
-                }
-                stop = stop || !(args.result_fn)(resp.contents);
-            }
-        }
-
-        Ok(())
     }
 
     pub async fn make_bucket(
